hooks: pull hooks into their own file. add memcpy hook.
closes #102.
